Fix some links
charlotteis.co.uk seems to have gone offline -- replace with links to
the Wayback Machine.

The Advanced Simulation and Computing link was redirecting to an article
about maintaining a nuclear stockpile.

Fix a link in episode 2 that was supposed to point back to episode 1 to
actually do so (instead of pointing to itself).